backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Werkzeug==2.3.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
